made difficulty row component to work with both highscore and settings
