+++ lib/app/app.dart
@@ -7,6 +7,8 @@ import 'package:template/theme/theme.dart';
 import 'package:provider/provider.dart';
 
 class App extends StatelessWidget {
+  const App({super.key});
+
   @override
   Widget build(BuildContext context) {
     return MultiProvider(
