+++ lib/components/difficulty_row.dart
@@ -2,11 +2,13 @@ import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 import 'package:template/components/custom_button.dart';
 import 'package:template/data/game_session.dart';
+import 'package:template/data/highscore.dart';
 import 'package:template/theme/theme.dart';
 
 class DifficultyRow extends StatelessWidget {
-  const DifficultyRow({super.key});
+  late bool isSettingsView = true;
 
+  DifficultyRow({super.key, required this.isSettingsView});
   @override
   Widget build(BuildContext context) {
     String difficulty =
@@ -48,13 +50,22 @@ class DifficultyRow extends StatelessWidget {
     return Opacity(
       opacity: opacity,
       child: CustomButton(
-        text: Text(newDifficulty, style: Themes.textStyle.headline3),
-        width: 80,
-        height: 40,
-        color: color,
-        onPressed: () => Provider.of<GameSession>(context, listen: false)
-            .updateDifficulty(newDifficulty.toLowerCase()),
-      ),
+          text: Text(newDifficulty, style: Themes.textStyle.headline3),
+          width: 80,
+          height: 40,
+          color: color,
+          onPressed: () {
+            if (isSettingsView) {
+              Provider.of<GameSession>(context, listen: false).updateDifficulty(
+                newDifficulty.toLowerCase(),
+              );
+            } else {
+              Provider.of<Highscore>(context, listen: false)
+                  .setDifficultyToView(newDifficulty.toLowerCase());
+              Provider.of<Highscore>(context, listen: false)
+                  .fetchScores(difficulty: newDifficulty.toLowerCase());
+            }
+          }),
     );
   }
 }
